Revert "update fragment location"
