+++ app/src/main/java/com/example/thefixer/ui/account/AccountFragment.java
@@ -1,23 +1,30 @@
-package com.example.thefixer;
+package com.example.thefixer.ui.account;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 
-public class HomeFragment extends Fragment {
+import com.example.thefixer.R;
+import com.example.thefixer.databinding.FragmentAccountBinding;
 
-    public HomeFragment(){
+public class AccountFragment extends Fragment {
+
+    public AccountFragment(){
 
     }
+
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view =  inflater.inflate(R.layout.fragment_home, container, false);
+        View view = inflater.inflate(R.layout.fragment_account, container, false);
         return view;
     }
