+++ app/src/main/java/com/example/thefixer/ui/activity/ActivityFragment.java
@@ -1,24 +1,27 @@
-package com.example.thefixer;
+package com.example.thefixer.ui.activity;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
+import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 
-public class PaymentFragment extends Fragment {
+import com.example.thefixer.R;
+import com.example.thefixer.databinding.FragmentActivityBinding;
 
-    public PaymentFragment(){
+public class ActivityFragment extends Fragment {
 
-    }
+    public ActivityFragment(){}
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_payment, container, false);
+        View view = inflater.inflate(R.layout.fragment_activity, container, false);
         return view;
     }
