+++ app/src/main/java/com/example/thefixer/ui/home/HomeFragment.java
@@ -1,30 +1,27 @@
-package com.example.thefixer;
+package com.example.thefixer.ui.home;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
-
 import com.example.thefixer.R;
-import com.example.thefixer.databinding.FragmentAccountBinding;
+import com.example.thefixer.databinding.FragmentHomeBinding;
 
-public class AccountFragment extends Fragment {
+public class HomeFragment extends Fragment {
 
-    public AccountFragment(){
+    public HomeFragment(){
 
     }
-
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_account, container, false);
+        View view =  inflater.inflate(R.layout.fragment_home, container, false);
         return view;
     }
