+++ app/src/main/java/com/example/thefixer/ui/inbox/InboxFragment.java
@@ -1,4 +1,4 @@
-package com.example.thefixer;
+package com.example.thefixer.ui.inbox;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
