+++ app/src/main/res/layout/fragment_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AccountFragment" >
+    tools:context=".ui.account.AccountFragment" >
 
     <TextView
         android:layout_width="match_parent"
