+++ app/src/main/res/layout/fragment_activity.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityFragment" >
+    tools:context=".ui.activity.ActivityFragment" >
 
     <TextView
         android:id="@+id/text_activity"
