+++ app/src/main/res/layout/fragment_home.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".HomeFragment">
+    tools:context=".ui.home.HomeFragment">
 
     <TableLayout
         android:layout_width="wrap_content"
@@ -26,8 +27,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:background="@drawable/rounded_corners"
-                    android:src="@drawable/electric_icon"
-                    android:contentDescription="@string/electric_icon"/>
+                    android:src="@drawable/electric_icon" />
 
                 <TextView
                     android:layout_width="match_parent"
@@ -35,7 +35,7 @@
                     android:gravity="center_horizontal"
                     android:text="Electronic"
                     android:textColor="@color/black"
-                    android:textSize="13sp"
+                    android:textSize="12dp"
                     android:textStyle="bold" />
             </LinearLayout>
 
@@ -57,7 +57,7 @@
                     android:gravity="center_horizontal"
                     android:text="Vehicle"
                     android:textColor="@color/black"
-                    android:textSize="13sp"
+                    android:textSize="12dp"
                     android:textStyle="bold" />
             </LinearLayout>
 
@@ -79,7 +79,7 @@
                     android:gravity="center_horizontal"
                     android:text="Computer"
                     android:textColor="@color/black"
-                    android:textSize="13sp"
+                    android:textSize="12dp"
                     android:textStyle="bold" />
             </LinearLayout>
 
@@ -107,7 +107,7 @@
                     android:gravity="center_horizontal"
                     android:text="Security"
                     android:textColor="@color/black"
-                    android:textSize="13sp"
+                    android:textSize="12dp"
                     android:textStyle="bold" />
             </LinearLayout>
 
@@ -129,7 +129,7 @@
                     android:gravity="center_horizontal"
                     android:text="Plumber"
                     android:textColor="@color/black"
-                    android:textSize="13sp"
+                    android:textSize="12dp"
                     android:textStyle="bold" />
             </LinearLayout>
 
@@ -151,7 +151,7 @@
                     android:gravity="center_horizontal"
                     android:text="Mobile"
                     android:textColor="@color/black"
-                    android:textSize="13sp"
+                    android:textSize="12dp"
                     android:textStyle="bold" />
             </LinearLayout>
         </TableRow>
@@ -165,8 +165,8 @@
         android:paddingVertical="10dp"
         android:text="Best Fixer Nearby"
         android:textColor="@color/black"
-        android:textSize="18sp"
+        android:textSize="18dp"
         android:textStyle="bold" />
 
-
+    
 </LinearLayout>
