+++ app/src/main/res/layout/fragment_inbox.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InboxFragment" >
+    tools:context=".ui.inbox.InboxFragment" >
 
     <TextView
         android:id="@+id/text_inbox"
