+++ app/src/main/res/layout/fragment_payment.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PaymentFragment" >
+    tools:context=".ui.payment.PaymentFragment" >
 
     <TextView
         android:id="@+id/text_payment"
