+++ app/src/main/res/values/strings.xml
@@ -5,5 +5,4 @@
     <string name="title_payment">Payment</string>
     <string name="title_inbox">Inbox</string>
     <string name="title_account">Account</string>
-    <string name="electric_icon">ContentDescription</string>
 </resources>
